Arel::Nodes::NotBetweenSymmetric
