+++ lib/arel/extensions.rb
@@ -33,6 +33,7 @@ require 'arel/extensions/similar'
 require 'arel/extensions/not_similar'
 require 'arel/extensions/not_between'
 require 'arel/extensions/between_symmetric'
+require 'arel/extensions/not_between_symmetric'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
