+++ lib/arel/extensions/nodes.rb
@@ -4,11 +4,6 @@
 
 module Arel
   module Nodes
-
-    # postgres only: https://www.postgresql.org/docs/9.1/functions-comparison.html
-    class NotBetweenSymmetric < Arel::Nodes::BetweenSymmetric
-    end
-
     Arel::Nodes::Function.class_eval do
       # postgres only: https://www.postgresql.org/docs/9.5/functions-aggregate.html
       attr_accessor :orders
